repo-updater: Remove old scheduler
This commit removes the old scheduler from repo-updater.

Closes #2063
