+++ schema/schema.go
@@ -133,8 +133,7 @@ type Discussions struct {
 
 // ExperimentalFeatures description: Experimental features to enable or disable. Features that are now enabled by default are marked as deprecated.
 type ExperimentalFeatures struct {
-	Discussions      string `json:"discussions,omitempty"`
-	UpdateScheduler2 string `json:"updateScheduler2,omitempty"`
+	Discussions string `json:"discussions,omitempty"`
 }
 
 // Extensions description: Configures Sourcegraph extensions.
