+++ schema/site.schema.json
@@ -32,12 +32,6 @@
           "type": "string",
           "enum": ["enabled", "disabled"],
           "default": "disabled"
-        },
-        "updateScheduler2": {
-          "description": "Enables a new update scheduler algorithm",
-          "type": "string",
-          "enum": ["enabled", "disabled"],
-          "default": "disabled"
         }
       },
       "group": "Experimental",
