+++ schema/site_stringdata.go
@@ -37,12 +37,6 @@ const SiteSchemaJSON = `{
           "type": "string",
           "enum": ["enabled", "disabled"],
           "default": "disabled"
-        },
-        "updateScheduler2": {
-          "description": "Enables a new update scheduler algorithm",
-          "type": "string",
-          "enum": ["enabled", "disabled"],
-          "default": "disabled"
         }
       },
       "group": "Experimental",
